pyproject.toml
[project]
name = "stock-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "yfinance>=1.0",
]
